reorganize and cleanup Symtab.h
- fix incomplete type error of a std::vector<relocationEntry>
  member declared before relocationEntry was defined

- move class ExceptionBlock to its own .h and .C

- move class relocationEntry to its own .h and .C

- move Symbol:: methods from Symtab.C to Symbol.C
