+++ symtabAPI/src/Symtab.C
@@ -2139,279 +2139,6 @@ SYMTAB_EXPORT LookupInterface::~LookupInterface()
 {
 }
 
-
-SYMTAB_EXPORT ExceptionBlock::ExceptionBlock(Offset tStart, 
-      unsigned tSize, 
-      Offset cStart) 
-: tryStart_(tStart), trySize_(tSize), catchStart_(cStart), hasTry_(true),
-  tryStart_ptr(0), tryEnd_ptr(0), catchStart_ptr(0), fdeStart_ptr(0), fdeEnd_ptr(0)
-{
-}
-
-   SYMTAB_EXPORT ExceptionBlock::ExceptionBlock(Offset cStart) 
-: tryStart_(0), trySize_(0), catchStart_(cStart), hasTry_(false),
-  tryStart_ptr(0), tryEnd_ptr(0), catchStart_ptr(0), fdeStart_ptr(0), fdeEnd_ptr(0)
-{
-}
-
-SYMTAB_EXPORT bool ExceptionBlock::hasTry() const
-{ 
-   return hasTry_; 
-}
-
-SYMTAB_EXPORT Offset ExceptionBlock::tryStart() const
-{ 
-   return tryStart_; 
-}
-
-SYMTAB_EXPORT Offset ExceptionBlock::tryEnd() const
-{ 
-   return tryStart_ + trySize_; 
-}
-
-SYMTAB_EXPORT Offset ExceptionBlock::trySize() const
-{
-   return trySize_; 
-}
-
-SYMTAB_EXPORT bool ExceptionBlock::contains(Offset a) const
-{ 
-   return (a >= tryStart_ && a < tryStart_ + trySize_); 
-}
-
-SYMTAB_EXPORT relocationEntry::relocationEntry() :
-   target_addr_(0), 
-   rel_addr_(0), 
-   addend_(0), 
-   rtype_(Region::RT_REL), 
-   name_(""), 
-   dynref_(NULL), 
-   relType_(0),
-   rel_struct_addr_(0)
-{
-}   
-
-SYMTAB_EXPORT relocationEntry::relocationEntry(Offset ta, Offset ra, std::string n, 
-      Symbol *dynref, unsigned long relType) :
-   target_addr_(ta), 
-   rel_addr_(ra), 
-   addend_(0), 
-   rtype_(Region::RT_REL), 
-   name_(n), 
-   dynref_(dynref), 
-   relType_(relType),
-   rel_struct_addr_(0)
-{
-}
-
-SYMTAB_EXPORT relocationEntry::relocationEntry(Offset ta, Offset ra, Offset add, 
-      std::string n, Symbol *dynref, unsigned long relType) :
-   target_addr_(ta), 
-   rel_addr_(ra), 
-   addend_(add), 
-   rtype_(Region::RT_REL), 
-   name_(n), 
-   dynref_(dynref), 
-   relType_(relType),
-   rel_struct_addr_(0)
-{
-}
-
-SYMTAB_EXPORT relocationEntry::relocationEntry(Offset ra, std::string n, 
-      Symbol *dynref, unsigned long relType, Region::RegionType rtype) :
-   target_addr_(0), 
-   rel_addr_(ra), 
-   addend_(0), 
-   rtype_(rtype), 
-   name_(n), 
-   dynref_(dynref), 
-   relType_(relType),
-   rel_struct_addr_(0)
-{
-}
-
-SYMTAB_EXPORT relocationEntry::relocationEntry(Offset ta, Offset ra, Offset add,
-        std::string n, Symbol *dynref, unsigned long relType,
-        Region::RegionType rtype) :
-    target_addr_(ta),
-    rel_addr_(ra),
-    addend_(add),
-    rtype_(rtype),
-    name_(n),
-    dynref_(dynref),
-    relType_(relType),
-    rel_struct_addr_(0)
-{
-}
-
-SYMTAB_EXPORT Offset relocationEntry::target_addr() const 
-{
-    return target_addr_;
-}
-
-SYMTAB_EXPORT void relocationEntry::setTargetAddr(const Offset off)
-{
-    target_addr_ = off;
-}
-
-SYMTAB_EXPORT Offset relocationEntry::rel_addr() const 
-{
-    return rel_addr_;
-}
-
-SYMTAB_EXPORT void relocationEntry::setRelAddr(const Offset value)
-{
-    rel_addr_ = value;
-}
-
-SYMTAB_EXPORT const string &relocationEntry::name() const 
-{
-    return name_;
-}
-
-SYMTAB_EXPORT Symbol *relocationEntry::getDynSym() const 
-{
-    return dynref_;
-}
-
-SYMTAB_EXPORT bool relocationEntry::addDynSym(Symbol *dynref) 
-{
-    dynref_ = dynref;
-    return true;
-}
-
-SYMTAB_EXPORT Region::RegionType relocationEntry::regionType() const
-{
-	return rtype_;
-}
-
-SYMTAB_EXPORT unsigned long relocationEntry::getRelType() const 
-{
-    return relType_;
-}
-
-SYMTAB_EXPORT Offset relocationEntry::addend() const
-{
-        return addend_;
-}
-
-SYMTAB_EXPORT void relocationEntry::setAddend(const Offset value)
-{
-        addend_ = value;
-}
-
-SYMTAB_EXPORT void relocationEntry::setRegionType(const Region::RegionType value)
-{
-        rtype_ = value;
-}
-
-SYMTAB_EXPORT void relocationEntry::setName(const std::string &newName) {
-    name_ = newName;
-}
-
-bool relocationEntry::operator==(const relocationEntry &r) const
-{
-	if (target_addr_ != r.target_addr_) return false;
-	if (rel_addr_ != r.rel_addr_) return false;
-	if (addend_ != r.addend_) return false;
-	if (rtype_ != r.rtype_) return false;
-	if (name_ != r.name_) return false;
-	if (relType_ != r.relType_) return false;
-	if (dynref_ && !r.dynref_) return false;
-	if (!dynref_ && r.dynref_) return false;
-	if (dynref_)
-	{
-		if (dynref_->getMangledName() != r.dynref_->getMangledName()) return false;
-		if (dynref_->getOffset() != r.dynref_->getOffset()) return false;
-	}
-
-	return true;
-}
-
-ostream & Dyninst::SymtabAPI::operator<< (ostream &os, const relocationEntry &r) 
-{
-    if( r.getDynSym() != NULL ) {
-        os << "Name: " << setw(20) << ( "'" + r.getDynSym()->getMangledName() + "'" );
-    }else{
-        os << "Name: " << setw(20) << r.name();
-    }
-    os << " Offset: " << std::hex << std::setfill('0') << setw(8) << r.rel_addr() 
-       << std::dec << std::setfill(' ')
-       << " Offset: " << std::hex << std::setfill('0') << setw(8) << r.target_addr() 
-       << std::dec << std::setfill(' ')
-       << " Addend: " << r.addend()
-       << " Region: " << Region::regionType2Str(r.regionType())
-       << " Type: " << setw(15) << relocationEntry::relType2Str(r.getRelType())
-       << "(" << r.getRelType() << ")";
-    if( r.getDynSym() != NULL ) {
-        os << " Symbol Offset: " << std::hex << std::setfill('0') << setw(8) << r.getDynSym()->getOffset();
-        os << std::dec << std::setfill(' ');
-        if( r.getDynSym()->isCommonStorage() ) {
-            os << " COM";
-        }else if( r.getDynSym()->getRegion() == NULL ) {
-            os << " UND";
-        }
-    }
-    return os;
-}
-
-const char *Symbol::symbolType2Str(SymbolType t) 
-{
-   switch (t) 
-   {
-      CASE_RETURN_STR(ST_UNKNOWN);
-      CASE_RETURN_STR(ST_FUNCTION);
-      CASE_RETURN_STR(ST_OBJECT);
-      CASE_RETURN_STR(ST_MODULE);
-      CASE_RETURN_STR(ST_SECTION);
-      CASE_RETURN_STR(ST_TLS);
-      CASE_RETURN_STR(ST_DELETED);
-      CASE_RETURN_STR(ST_NOTYPE);
-      CASE_RETURN_STR(ST_INDIRECT);
-   };
-
-   return "invalid symbol type";
-}
-
-const char *Symbol::symbolLinkage2Str(SymbolLinkage t) 
-{
-   switch (t) 
-   {
-      CASE_RETURN_STR(SL_UNKNOWN);
-      CASE_RETURN_STR(SL_GLOBAL);
-      CASE_RETURN_STR(SL_LOCAL);
-      CASE_RETURN_STR(SL_WEAK);
-      CASE_RETURN_STR(SL_UNIQUE);
-   };
-
-   return "invalid symbol linkage";
-}
-
-const char *Symbol::symbolTag2Str(SymbolTag t) 
-{
-   switch (t) 
-   {
-      CASE_RETURN_STR(TAG_UNKNOWN);
-      CASE_RETURN_STR(TAG_USER);
-      CASE_RETURN_STR(TAG_LIBRARY);
-      CASE_RETURN_STR(TAG_INTERNAL);
-   };
-
-   return "invalid symbol tag";
-}
-
-const char *Symbol::symbolVisibility2Str(SymbolVisibility t) 
-{
-   switch(t) {
-      CASE_RETURN_STR(SV_UNKNOWN);
-      CASE_RETURN_STR(SV_DEFAULT);
-      CASE_RETURN_STR(SV_INTERNAL);
-      CASE_RETURN_STR(SV_HIDDEN);
-      CASE_RETURN_STR(SV_PROTECTED);
-   }
-   return "invalid symbol visibility";
-}
-
 bool Symtab::hasStackwalkDebugInfo()
 {
